Refactor batch api method
