+++ src/api-client/consts.ts
@@ -1,1 +1,2 @@
 export const BATCH_DUMP_TIMEOUT_MS = 2000;
+export const BATCH_URL = '/file-batch-api';
