+++ src/api-client/index.ts
@@ -1,2 +1,3 @@
 export * from './api-client';
 export * from './consts';
+export * from './files-api';
